add GameObjectMotion script
