--- Subject404/Assets/Scripts/AxeMotion.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0b5bc668265889d4783d58ddd30e5d85
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
