+++ Subject404/Assets/Scripts/CustomFunctionOnTrigger.cs
@@ -1,5 +1,4 @@
-using System.Collections;
-using System.Collections.Generic;
+using System;
 using UnityEngine;
 
 public class CustomFunctionOnTrigger : MonoBehaviour
