+++ Subject404/Assets/Scripts/GameObjectMotion.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0f68881dfd1acd64d82b7a3defc5a870
+guid: 65fb4bd26f4c6ff4c823afb731f84efa
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
